git-svn-id: svn://192.168.1.10/public/CoffeeMud@11230 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Common/DefaultQuest.java
@@ -2880,7 +2880,7 @@ public class DefaultQuest implements Quest, Tickable, CMObject
 		CMLib.quests().addQuest(Q2);
 		if(reTime)
 		{
-			Long ellapsed=(Long)stepEllapsedTimes.get(script);
+			Long ellapsed=stepEllapsedTimes.get(script);
 			if(ellapsed==null) ellapsed=Long.valueOf(0);
 			stepEllapsedTimes.remove(script);
 			ellapsed=Long.valueOf(ellapsed.longValue()+(System.currentTimeMillis()-lastStartDateTime));
